+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/extensions/IntelliJPlatformTestingExtension.kt
@@ -139,6 +139,7 @@ abstract class IntelliJPlatformTestingExtension @Inject constructor(private val 
                     intelliJPlatformConfiguration = customIntelliJPlatformConfiguration
                     intelliJPlatformPluginConfiguration = customIntellijPlatformPluginConfiguration
 
+                    sandboxDirectory = this@all.sandboxDirectory
                     splitMode = this@all.splitMode.orElse(basePrepareSandboxTask.flatMap { it.splitMode })
                     splitModeTarget = this@all.splitModeTarget.orElse(basePrepareSandboxTask.flatMap { it.splitModeTarget })
                     disabledPlugins = plugins.disabled